+++ library/src/test/java/org/mustangproject/ZUGFeRD/ZF2ZInvoiceImporterTest.java
@@ -550,6 +550,7 @@ public class ZF2ZInvoiceImporterTest extends ResourceCase {
 				"      \"totalAmount\" : 0.1,\n" +
 				"      \"taxRateApplicablePercent\" : 0,\n" +
 				"      \"taxCategoryCode\" : \"S\"\n" +
+				"      \"reasonCode\" : \"95\",\n" +
 				"    } ],\n" +
 				"    \"value\" : 3.0,\n" +
 				"    \"calculation\" : {\n" +
@@ -602,6 +603,7 @@ public class ZF2ZInvoiceImporterTest extends ResourceCase {
 				"      \"taxRateApplicablePercent\" : 0,\n" +
 				"      \"reason\" : \"AnotherReason\",\n" +
 				"      \"taxCategoryCode\" : \"S\"\n" +
+				"      \"reasonCode\" : \"ABK\",\n" +
 				"    } ],\n" +
 				"    \"value\" : 3.0,\n" +
 				"    \"calculation\" : {\n" +
@@ -633,6 +635,7 @@ public class ZF2ZInvoiceImporterTest extends ResourceCase {
 				"      \"taxRateApplicablePercent\" : 0,\n" +
 				"      \"reason\" : \"Yet another reason\",\n" +
 				"      \"taxCategoryCode\" : \"S\"\n" +
+				"      \"reasonCode\" : \"ABK\",\n" +
 				"    } ],\n" +
 				"    \"value\" : 3.0,\n" +
 				"    \"calculation\" : {\n" +
@@ -1012,5 +1015,21 @@ public class ZF2ZInvoiceImporterTest extends ResourceCase {
 			}
 		}
 		return null;
+	}
+
+	@Test
+	public void testRecalc() throws FileNotFoundException, XPathExpressionException, ParseException {
+		File inputFile = getResourceAsFile("XRechnung_internalRecalcBug.xml");
+		ZUGFeRDInvoiceImporter zii = new ZUGFeRDInvoiceImporter();
+		zii.doRecalculateItemPricesFromLineTotals();
+		// zii.doIgnoreCalculationErrors();
+		zii.setInputStream(new FileInputStream(inputFile));
+
+		CalculatedInvoice invoice = new CalculatedInvoice();
+		zii.extractInto(invoice);
+
+		assertEquals(new BigDecimal("0.012148"), invoice.getZFItems()[0].getPrice());
+		assertEquals(new BigDecimal("12.44"), invoice.getLineTotalAmount().setScale(2));
+		assertEquals(new BigDecimal("12.44"), invoice.getGrandTotal().setScale(2));
 	}
 }